pyproject.toml
[tool.poetry]
name = "auth-fastapi"
version = "0.1.0"
description = ""
authors = ["isinthesky <hijde@naver.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
ruff = "^0.8.2"
pytest = "^8.3.4"
icecream = "^2.1.3"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
kombu = "^5.4.2"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.10.3"}
asyncpg = "^0.30.0"
greenlet = "^3.1.1"

[tool.poetry.scripts]
start = "src.main:start"
version = "increment_version:start"
lint-strict = "ruff check . --select ALL"
lint-fix = "ruff check . --fix"
format-check = "ruff format . --check"
format-fix = "ruff format ."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
